- In debit entry link to academic event, if associated and not event annuled - On annulments of debit entries do not delete treasury events - Treasury events may be annuled when debit entry is not closed in debit note, but exemption reversion is only possible when debit entry is in preparation - In treasury event links to debit notes, credit notes and debit entries - Fix label about academic act block in read debit entry page, to be conformant with update debit entry form - On creating exemption show finantial document associated with debit entry - Only allow treasury exemption deletion when debit entry is annuled or is not associated with debit note or debit note is in preparing mode - But only allow treasury exemption deletion in interface if debit entry is in preparing and active in treasury event

+++ src/main/java/org/fenixedu/treasury/ui/accounting/managecustomer/TreasuryEventController.java
@@ -40,6 +40,7 @@ import org.fenixedu.treasury.domain.debt.DebtAccount;
 import org.fenixedu.treasury.domain.document.CreditEntry;
 import org.fenixedu.treasury.domain.document.DebitEntry;
 import org.fenixedu.treasury.domain.event.TreasuryEvent;
+import org.fenixedu.treasury.domain.exceptions.TreasuryDomainException;
 import org.fenixedu.treasury.domain.exemption.TreasuryExemption;
 import org.fenixedu.treasury.ui.TreasuryBaseController;
 import org.fenixedu.treasury.util.Constants;
@@ -215,7 +216,13 @@ public class TreasuryEventController extends TreasuryBaseController {
 
         try {
             assertUserIsFrontOfficeMember(model);
-
+            
+            if(treasuryExemption.getDebitEntry().isAnnulled() 
+                    || treasuryExemption.getDebitEntry().getTreasuryEvent() == null
+                    || treasuryExemption.getDebitEntry().isEventAnnuled()) {
+                throw new TreasuryDomainException("error.TreasuryExemption.delete.impossible");
+            }
+            
             treasuryExemption.delete();
 
             addInfoMessage(BundleUtil.getString(Constants.BUNDLE, "label.success.delete"), model);
